requirements.txt
langchain
langchain-community
langchain-openai
streamlit
pypdf
chromadb
python-dotenv
tiktoken
faiss-cpu